Cleanup dates formatting
