+++ content/projects/covid-tracker.mdx
@@ -1,6 +1,6 @@
 ---
 title: COVID Tracker
-gist: A web app that quickly gives you COVID-19 data. Best Halloween Hack of TCNJ HackDown Fall 2020!
+gist: Web app that quickly gives you COVID-19 data. Won Best Halloween Hack of TCNJ HackDown Fall 2020!
 start: October 2020
 end: November 2020
 repo: https://github.com/samuel-ping/covid-tracker
