+++ content/projects/food-mood.mdx
@@ -1,7 +1,6 @@
 ---
 title: Food Mood
-gist: Web app that takes your photo, detects your visible mood, and
-  suggests a restaurant based off that mood!
+gist: Web app that suggests a restaurant based off your mood!
 start: March 2020
 end: August 2020
 repo: https://github.com/samuel-ping/food-mood
