+++ content/projects/personal-website.mdx
@@ -1,6 +1,6 @@
 ---
 title: Personal Website
-gist: My personal website!
+gist: The site you're looking at right now!
 start: May 2020
 end: ''
 repo: https://github.com/samuel-ping/personal-website
