+++ content/projects/plantfam.mdx
@@ -1,6 +1,6 @@
 ---
 title: PlantFam
-gist: Plant tracking mobile application.
+gist: Android app that helps you track the progress of your plants’ growth.
 start: January 2022
 end: ''
 repo: https://github.com/samuel-ping/PlantFam-Android
