+++ content/projects/street-cred.mdx
@@ -1,7 +1,6 @@
 ---
 title: Street Cred
-gist: Web app that calculates your credit score and generates
-  feedback based on the five main categories that make up your credit score.
+gist: Web app that calculates and gives feedback on your credit score.
 start: August 2020
 end: August 2020
 repo: https://github.com/samuel-ping/street-cred
